1st Com

+++ src/main/java/uz/mamatsalay/incomemanagementservice/repository/UserRepository.java
@@ -1,16 +1,13 @@
 package uz.mamatsalay.incomemanagementservice.repository;
 
+import uz.mamatsalay.incomemanagementservice.model.User;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import uz.mamatsalay.incomemanagementservice.model.User;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends CrudRepository<User, Integer> {
-
+public interface UserRepository extends CrudRepository<User, Long> {
     Optional<User> findByEmail(String email);
-
     Optional<User> findByVerificationCode(String verificationCode);
-
 }
